update member guidelines with EU sponsor asks
MLH is constantly working with more sponsors to be able to support the EU hackathon community and would like to make these benefits and requirements clearer to current and prospective hackathon organizers.  Being an EU Member Event with MLH may means these events will have more direct access to sponsors in-person for their event in exchange for some brand presence at the event.