2026 Season Updates (#243)
* Update README.md

Add additional pages that were accidentally left out when they were originally added

* Update community-values.md

minor updates based on review with Ryan and Mary

* Update disciplinary-action.md

Updates based on review with Ryan and Mary

* Update 3rd checkbox participant-notices-and-agreements.md

This 3rd checkbox update was actually done a year or 2 ago? It seems it was missed when it was updated on the documentation sent to organizers.

* Update incident-response.md